Move system messages init into separate module (#4251)
* Move system messages init into separate module

* getSystemMessages => initSystemMessages
